--- Assets/MainMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 078e86d323af34c438f7872ebc7b9f9b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
