--- Assets/Plugins.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 720712be6e31c444f88f57bb85bf5fa5
-folderAsset: yes
-timeCreated: 1504268238
-licenseType: Store
-DefaultImporter:
-  userData:
-  assetBundleName:
-  assetBundleVariant:
