+++ Assets/Scenes/GameView.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a7d5d0ba26b66174ca7f9e6a0da6b8b7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
