+++ Assets/Scenes/GameView/Materials/BlackSquare.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7fccd66520c8fe84cae85049c9c6db26
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
