+++ Assets/Scenes/GameView/Materials/WhiteSquare.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 958e3843a39aaa1428f17ca4e79eedaa
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
