+++ Assets/Scenes/Menu/Art.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f3390d3773e6fa64ca6b9a4bdd685864
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
