+++ Assets/Scenes/Menu/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b132568314e01244b800269aa50cb754
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
