+++ Assets/Scenes/Menu/TextGradient.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9b086a72b13c5c64b9985ad9596e6c40
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
