--- Assets/TextGradient.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9b086a72b13c5c64b9985ad9596e6c40
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
